alx_travel_app/requirement.txt
django
djangorestframework
django-cors-headers
celery
rabbitmq
drf-yasg
python-environ
mysqlclient
mysql-connector-python
python-dotenv
django-environ
pycodestyle
pylint
drf-nested-routers